Changing the directory structure to account for multiple implementations
