+++ deployment/docker/etc/org.codice.ddf.security.idp.server.IdpEndpoint.config
