+++ deployment/docker/etc/org.codice.ddf.security.policy.context.impl.PolicyManager.config
