--- distribution/suites/build.gradle
@@ -1,7 +1,0 @@
-group = 'org.codice.samlconf.distribution'
-
-description = 'Suites to run tests.'
-
-dependencies {
-    compile project(':test:idp')
-}
