+++ settings.gradle
@@ -1,13 +1,12 @@
 rootProject.name = 'saml-conformance'
 include ':library'
-include ':plugins:plugins-api'
-include ':plugins:ddf-plugins'
-include ':plugins'
+include ':external:api'
+include ':external:implementations:ddf'
+include ':external:implementations:keycloak'
 include ':test:common'
 include ':test:idp'
 include ':test:sp'
-include ':test'
-include ':distribution:command-line'
-include ':distribution:suites'
-include ':distribution:docker'
-include ':distribution'
+include ':deployment:distribution'
+include ':deployment:suites'
+include ':deployment:docker'
+
