+++ test/common/build.gradle
@@ -4,5 +4,5 @@ description = 'Common Functions for IdP and SP Tests.'
 dependencies {
     compile 'io.kotlintest:kotlintest:2.0.7'
     compile project(':library')
-    compile project(':plugins:plugins-api')
+    compile project(':external:api')
 }
